se mueve logica de obtener preguntas de clase al comp RespuestasAPreguntas
